refact: Car의 name 스트링을 Name 클래스로 wrapping
