+++ src/main/java/racingcar/domain/Name.java
@@ -1,0 +1,40 @@
+package racingcar.domain;
+
+import racingcar.exception.InvalidNameLengthException;
+
+import java.util.Objects;
+
+public class Name {
+
+    private static final int MAX_CAR_NAME_LENGTH = 5;
+
+    private final String name;
+
+    public Name(String name) {
+        validateNameLength(name);
+        this.name = name;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    private void validateNameLength(String name) {
+        if (name.length() > MAX_CAR_NAME_LENGTH) {
+            throw new InvalidNameLengthException("이름은 5자 이하만 가능합니다.");
+        }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Name name1 = (Name) o;
+        return Objects.equals(name, name1.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+}
